rds-to-s3-using-lambda-aws-sdk-for-pandas/code/requirements.txt
psycopg2-binary==2.9.5
python-dateutil==2.8.2
six==1.16.0
